requirements.txt
# API Dependencies
flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1

# Data Processing
elasticsearch==8.12.1
kafka-python==2.0.2
pandas==2.2.1
numpy==1.26.4

# NLP and Text Processing
spacy==3.7.4
geopy==2.4.1
geotext==0.4.0
pycountry==23.12.11

# Utilities
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
urllib3==2.2.1

# Development Dependencies
pytest==8.0.2
black==24.2.0
flake8==7.0.0
mypy==1.8.0
